requirements.txt
openstacksdk
icecream
click